editoast: migrate rocket to actix

+++ editoast/src/client/postgres_config.rs
@@ -21,7 +21,7 @@ pub struct PostgresConfig {
     pub psql_port: u16,
     #[derivative(Default(value = "32"))]
     #[clap(long, env, default_value_t = 32)]
-    pub pool_size: u16,
+    pub pool_size: u32,
 }
 
 impl PostgresConfig {
